==> Step1-DataCollection.ipynb <==
## Twitter Data Collection

```Tweepy``` was used to retrieve tweets with the keywords, tax, #taxday, and taxation, to analyze the overall sentiment around taxes.  

# Import required libraries
import tweepy
import json
import pprint
import pickle

# Set up Authentication to Twitter
con_key = '#'
con_secret = '#'
acc_token = '#'
acc_secret = '#'

# Use tweepy.OAuthHandler to create an authentication using the given key and secret
auth = tweepy.OAuthHandler(consumer_key=con_key, consumer_secret=con_secret)
auth.set_access_token(acc_token, acc_secret)

# Connect to the Twitter API using the authentication
api = tweepy.API(auth, wait_on_rate_limit = True, wait_on_rate_limit_notify = True)

The Streaming API was used to retrieve live tweets on Monday, 4/18/16, which was the last day to complete taxes without penalty.

# Create a subclass of tweepy.StreamListener to add a response to on_status
class PrintingStreamListener(tweepy.StreamListener):

    def on_status(self, status):
        print(status.text)
        
    # Save streaming tweets to a JSON file for analysis later
    def on_data(self, data):
        try:
            with open('tweets.json', 'a') as f:
                f.write(data)
                return True
        except BaseException as e:
            print("Error on_data: %s" % str(e))
        return True
        
    #disconnect the stream if we receive an error message indicating we are overloading Twitter
    def on_error(self, status_code):
        if status_code == 420:
            #returning False in on_data disconnects the stream
            return False

# Create and authenticate an instance of our new ```PrintingStreamListener``` class
my_stream_listener = PrintingStreamListener()
my_stream = tweepy.Stream(auth = api.auth, listener=my_stream_listener)

# Look for recent tweets with the keywords, tax, taxation, taxday.
my_stream.filter(track=['tax', 'taxation', 'taxday'])

# Disconnect from twitter stream
my_stream.disconnect()

To classify the twitter users's political affiliation, a list of followers for each of the 4 major presidential condidates was dowloaded using ```tweepy```.

# Get a list of Donald Trump followers
donald_trump_followers = []
for user in tweepy.Cursor(api.followers, screen_name="realDonaldTrump").items():
    donald_trump_followers.append(user)

# Get a list of Ted Cruz followers
ted_cruz_followers = []
for user in tweepy.Cursor(api.followers, screen_name="tedcruz").items():
    ted_cruz_followers.append(user)

# Get a list of Bernie Sanders followers
bernie_sanders_followers = []
for user in tweepy.Cursor(api.followers, screen_name="BernieSanders").items():
    bernie_sanders_followers.append(user)

# Get a list of Hilary Cliton followers
hilary_clinton_followers = []
for user in tweepy.Cursor(api.followers, screen_name="HilaryClinton").items():
    hilary_clinton_followers.append(user)

# Save the list of trump followers as a pickle file for analysis later
with open("trump2.pickle","wb") as f:
    pickle.dump(trump_list, f)

# Save the list of cruz followers as a pickle file for analysis later
with open("cruz.pickle","wb") as f:
    pickle.dump(ted_cruz_followers, f)

# Save the list of bernie followers as a pickle file for analysis later
with open("bernie.pickle","wb") as f:
    pickle.dump(bernie_sanders_followers, f)

# Save the list of trump followers as a pickle file for analysis later
with open("hilary.pickle","wb") as f:
    pickle.dump(hilary_clinton_followers, f)
